Generate expand query property for OData 4 in byKey method (fix T670392)
