+++ js/data/odata/store.js
@@ -1,5 +1,4 @@
 var isDefined = require("../../core/utils/type").isDefined,
-    map = require("../../core/utils/iterator").map,
     odataUtils = require("./utils"),
     proxyUrlFormatter = require("../proxy_url_formatter"),
     errors = require("../errors").errors,
@@ -160,7 +159,7 @@ var ODataStore = Store.inherit({
 
         if(extraOptions) {
             if(extraOptions.expand) {
-                params["$expand"] = map([].slice.call(extraOptions.expand), odataUtils.serializePropName).join();
+                params["$expand"] = odataUtils.generateExpand(this._version, extraOptions.expand);
             }
         }
