+++ js/data/odata/utils.js
@@ -1,10 +1,13 @@
 var Class = require("../../core/class"),
     extend = require("../../core/utils/extend").extend,
     typeUtils = require("../../core/utils/type"),
+    iteratorUtils = require("../../core/utils/iterator"),
     each = require("../../core/utils/iterator").each,
     ajax = require("../../core/utils/ajax"),
     Guid = require("../../core/guid"),
     isDefined = typeUtils.isDefined,
+    isPlainObject = typeUtils.isPlainObject,
+    grep = require("../../core/utils/common").grep,
     Deferred = require("../../core/utils/deferred").Deferred,
 
     errors = require("../errors").errors,
@@ -17,6 +20,14 @@ var ISO8601_DATE_REGEX = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[-+]{1}
 
 // Request processing
 var JSON_VERBOSE_MIME_TYPE = "application/json;odata=verbose";
+
+var makeArray = function(value) {
+    return typeUtils.type(value) === "string" ? value.split() : value;
+};
+
+var hasDot = function(x) {
+    return /\./.test(x);
+};
 
 function formatISO8601(date, skipZeroTime, skipTimezone) {
     var bag = [];
@@ -317,7 +328,7 @@ var interpretJsonFormat = function(obj, textStatus, transformOptions) {
         return { error: error };
     }
 
-    if(!typeUtils.isPlainObject(obj)) {
+    if(!isPlainObject(obj)) {
         return { data: obj };
     }
 
@@ -483,7 +494,7 @@ var serializeValue = function(value, protocolVersion) {
 };
 
 var serializeKey = function(key, protocolVersion) {
-    if(typeUtils.isPlainObject(key)) {
+    if(isPlainObject(key)) {
         var parts = [];
         each(key, function(k, v) {
             parts.push(serializePropName(k) + "=" + serializeValue(v, protocolVersion));
@@ -553,12 +564,150 @@ var convertPrimitiveValue = function(type, value) {
     return converter(value);
 };
 
+var generateSelect = function(oDataVersion, select) {
+    if(!select) {
+        return;
+    }
+
+    if(oDataVersion < 4) {
+        return serializePropName(select.join());
+    }
+
+    return grep(select, hasDot, true).join();
+};
+
+var generateExpand = function(oDataVersion, expand, select) {
+    var generatorV2 = function() {
+        var hash = {};
+
+        if(expand) {
+            iteratorUtils.each(makeArray(expand), function() {
+                hash[serializePropName(this)] = 1;
+            });
+        }
+
+        if(select) {
+            iteratorUtils.each(makeArray(select), function() {
+                var path = this.split(".");
+                if(path.length < 2) {
+                    return;
+                }
+
+                path.pop();
+                hash[serializePropName(path.join("."))] = 1;
+            });
+        }
+
+        return iteratorUtils.map(hash, function(k, v) { return v; }).join();
+    };
+
+    var generatorV4 = function() {
+        var format = function(hash) {
+            var formatCore = function(hash) {
+                var result = "",
+                    selectValue = [],
+                    expandValue = [];
+
+                iteratorUtils.each(hash, function(key, value) {
+                    if(Array.isArray(value)) {
+                        [].push.apply(selectValue, value);
+                    }
+
+                    if(isPlainObject(value)) {
+                        expandValue.push(key + formatCore(value));
+                    }
+                });
+
+                if(selectValue.length || expandValue.length) {
+                    result += "(";
+
+                    if(selectValue.length) {
+                        result += "$select=" + iteratorUtils.map(selectValue, serializePropName).join();
+                    }
+
+                    if(expandValue.length) {
+                        if(selectValue.length) {
+                            result += ";";
+                        }
+
+                        result += "$expand=" + iteratorUtils.map(expandValue, serializePropName).join();
+                    }
+                    result += ")";
+                }
+
+                return result;
+            };
+
+            var result = [];
+
+            iteratorUtils.each(hash, function(key, value) {
+                result.push(key + formatCore(value));
+            });
+
+            return result.join();
+        };
+
+        var parseTree = function(exprs, root, stepper) {
+            var parseCore = function(exprParts, root, stepper) {
+                var result = stepper(root, exprParts.shift(), exprParts);
+                if(result === false) {
+                    return;
+                }
+
+                parseCore(exprParts, result, stepper);
+            };
+
+            iteratorUtils.each(exprs, function(_, x) {
+                parseCore(x.split("."), root, stepper);
+            });
+        };
+
+        var hash = {};
+
+        if(expand || select) {
+            if(expand) {
+                parseTree(makeArray(expand), hash, function(node, key, path) {
+                    node[key] = node[key] || {};
+
+                    if(!path.length) {
+                        return false;
+                    }
+
+                    return node[key];
+                });
+            }
+
+            if(select) {
+                parseTree(grep(makeArray(select), hasDot), hash, function(node, key, path) {
+                    if(!path.length) {
+                        node[key] = node[key] || [];
+                        node[key].push(key);
+                        return false;
+                    }
+
+                    return (node[key] = node[key] || {});
+                });
+            }
+
+            return format(hash);
+        }
+    };
+
+    if(oDataVersion < 4) {
+        return generatorV2();
+    }
+
+    return generatorV4();
+};
+
 exports.sendRequest = sendRequest;
 exports.serializePropName = serializePropName;
 exports.serializeValue = serializeValue;
 exports.serializeKey = serializeKey;
 exports.keyConverters = keyConverters;
 exports.convertPrimitiveValue = convertPrimitiveValue;
+exports.generateExpand = generateExpand;
+exports.generateSelect = generateSelect;
 
 exports.EdmLiteral = EdmLiteral;
